back_end/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
mongoengine==0.27.0
pymongo==4.6.0
setuptools==69.0.0
dnspython==2.4.2